MOTECH-1913: IT tests for CourseStructureService
