+++ mtraining/src/test/java/org/motechproject/mtraining/service/it/MTrainingIntegrationTests.java
@@ -7,7 +7,7 @@ import org.junit.runners.Suite;
  * Parent IT class to run all the individual service ITs
  */
 @RunWith(Suite.class)
-@Suite.SuiteClasses({MTrainingServiceBundleIT.class, CourseStructureServiceIT.class,
+@Suite.SuiteClasses({MTrainingServiceBundleIT.class, CourseStructureServiceBundleIT.class,
         BookmarkServiceBundleIT.class, ActivityServiceBundleIT.class})
 public class MTrainingIntegrationTests {
 }
